fix(platform): review fixes for task 7 — semantics doc + copyEvent hoist
Addresses code-review findings on the WI-2 char-span-boundary pin:
- Updates the clipboard semantics doc's S6 section and test-mapping table
  with the outcome (does not reproduce at the Lexical selection level;
  root cause narrowed to the DOM Range resolution layer, still covered
  only by Task 8's E2E steps).
- Hoists the fourth duplicate `copyEvent()` clipboard stub (added by this
  task's own commit) into markerEdit.test-helpers.tsx and updates all four
  call sites to import the shared helper. No behavior change.
