+++ packages/platform/src/editor/markerEdit/markerEditDeletion.utils.test.tsx
@@ -6,6 +6,7 @@ import {
 import {
   $appendCharPara,
   $appendVersePara,
+  copyEvent,
   requireDefined,
   serializedState,
   testEnvironment,
@@ -1031,25 +1032,6 @@ describe("selection-delete at a settled char-span boundary (WI-2 filed)", () => 
       serializedState(quoteWjUsj()),
       <MarkerEditPlugin viewOptions={standardViewOptions} />,
     );
-  }
-
-  /**
-   * jsdom (see `whitespaceDisplay.plugin.utils.test.tsx`'s identically-shaped stub) doesn't
-   * implement `ClipboardEvent`/`DataTransfer`; the cut handler under test only touches
-   * `clipboardData.getData`/`setData`/`preventDefault`, so a minimal stub covers dispatch.
-   */
-  function copyEvent(): { event: ClipboardEvent; getData: (type: string) => string } {
-    const store = new Map<string, string>();
-    const clipboardData = {
-      getData: (type: string) => store.get(type) ?? "",
-      setData: (type: string, data: string) => {
-        store.set(type, data);
-      },
-    };
-    return {
-      event: { clipboardData, preventDefault: vi.fn() } as unknown as ClipboardEvent,
-      getData: (type: string) => clipboardData.getData(type),
-    };
   }
 
   /**
